Convert prediction to float (#78)
* rewards: convert predictions_path to float

* fix error message in create_database_engine

* Update synth/validator/reward.py

+++ synth/validator/reward.py
@@ -83,7 +83,7 @@ def reward(
 
     try:
         score, detailed_crps_data = calculate_crps_for_miner(
-            np.array(predictions_path),
+            np.array(predictions_path).astype(float),
             np.array(real_price_path),
             simulation_input.time_increment,
         )
